+++ inc/Helpers/DisplayImplementation.php
@@ -54,6 +54,15 @@ class DisplayImplementation {
 					return $name;
 				}
 
+				 if ( is_string($value) ) {
+					return $name . '="' . esc_attr($value) . '"';
+				}
+
+				if ( is_callable( $value ) ) {
+					// phpcs:ignore NeutronStandard.Functions.VariableFunctions.VariableFunction
+					$value = $value();
+				}
+
 				if ( is_array( $value ) || is_object( $value ) ) {
 					$value = wp_json_encode( $value );
 				}